# docs/openapi.yaml
openapi: '3.0.3'
info:
    title: 'Carpooling'
    version: '0.2.4'
paths:
...
    /user/{id}:
        get:
            summary: 'Get a user'
            description: 'Get a user by their user id'
            tags:
                - user
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      $ref: 'definitions.yaml#/components/schemas/Id'
                  description: 'The user ID'
            responses:
                '200':
                    description: 'Returns the user'
                    content:
                        application/json:
                            schema:
                                $ref: 'definitions.yaml#/components/schemas/User'
                '400':
                    description: 'No id provided'
                '401':
                    description: 'Unauthenticated'
                '404':
                    description: 'User not found'
    /user/email/{email}:
        get:
            summary: 'Get a user by email'
            description: 'Get a user by their email'
            tags:
                - user
            parameters:
                - in: path
                  name: email
                  required: true
                  schema:
                      type: string
                  description: 'The user email'
            responses:
                '200':
                    description: 'Returns the user'
                    content:
                        application/json:
                            schema:
                                $ref: 'definitions.yaml#/components/schemas/User'
                '400':
                    description: 'No email provided'
                '401':
                    description: 'Unauthenticated'
                '404':
                    description: 'User not found'
    /user:
        patch:
            summary: 'Update a user'
            description: 'Update a user with new data, omitted fields are unchanged.'
            tags:
                - user
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      $ref: 'definitions.yaml#/components/schemas/Id'
                  description: 'The user ID'
            requestBody:
                description: 'The updated user object'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                    description: 'New email of the user'
                                friends:
                                    type: array
                                    description: 'New friend list of user'
                                    items:
                                        $ref: 'definitions.yaml#/components/schemas/Id'
                                avatar:
                                    type: string
                                    description: 'New avatar image of user'
            responses:
                '200':
                    description: 'Returns the updated user'
                    content:
                        application/json:
                            schema:
                                $ref: 'definitions.yaml#/components/schemas/User'
                '401':
                    description: 'Unauthenticated'
                '404':
                    description: 'User not found'

        post:
            summary: 'Create a user'
            description: 'Create a user from the provided data and firebase auth token'
            tags:
                - user
            requestBody:
                description: 'User details'
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                firstName:
                                    type: string
                                    description: 'First name of the user'
                                lastName:
                                    type: string
                                    description: 'Last name of the user'
                                dateOfBirth:
                                    type: number
                                    description: 'Date of birth of the user'
                                avatar:
                                    type: string
                                    description: 'User avatar image'
                            required:
                                - firstName
                                - lastName
                                - dateOfBirth
            responses:
                '200':
                    description: 'Returns the newly created user'
                    content:
                        application/json:
                            schema:
                                $ref: 'definitions.yaml#/components/schemas/User'
                '401':
                    description: 'Unauthenticated'
                '409':
                    description: 'User already exists'
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: FBAuth
security:
    - bearerAuth: []